pre-submit cleanup
